refactor: relocate enclose_response and enclose_response_raw to communication.py
