+++ python/src/uagents/query.py
@@ -5,9 +5,7 @@ from time import time
 from typing import Optional
 
 import aiohttp
-from pydantic import UUID4
 from uagents.crypto import generate_user_address
-from uagents.dispatch import JsonStr
 from uagents.envelope import Envelope
 from uagents.models import Model
 from uagents.resolver import GlobalResolver, Resolver
@@ -90,53 +88,3 @@ async def query(
             )
 
     LOGGER.exception(f"Failed to send sync message to {destination}")
-
-
-def enclose_response(
-    message: Model, sender: str, session: UUID4, target: str = ""
-) -> str:
-    """
-    Enclose a response message within an envelope.
-
-    Args:
-        message (Model): The response message to enclose.
-        sender (str): The sender's address.
-        session (str): The session identifier.
-        target (str): The target address.
-
-    Returns:
-        str: The JSON representation of the response envelope.
-    """
-    schema_digest = Model.build_schema_digest(message)
-    return enclose_response_raw(message.json(), schema_digest, sender, session, target)
-
-
-def enclose_response_raw(
-    json_message: JsonStr,
-    schema_digest: str,
-    sender: str,
-    session: UUID4,
-    target: str = "",
-) -> str:
-    """
-    Enclose a raw response message within an envelope.
-
-    Args:
-        json_message (JsonStr): The JSON-formatted response message to enclose.
-        schema_digest (str): The schema digest of the message.
-        sender (str): The sender's address.
-        session (UUID4): The session identifier.
-        target (str): The target address.
-
-    Returns:
-        str: The JSON representation of the response envelope.
-    """
-    response_env = Envelope(
-        version=1,
-        sender=sender,
-        target=target,
-        session=session,
-        schema_digest=schema_digest,
-    )
-    response_env.encode_payload(json_message)
-    return response_env.json()
